requirements.txt
django==5.1.7
psycopg2-binary==2.9.11
python-decouple==3.8
django-htmx==1.27.0
requests==2.32.3
pdfplumber
gunicorn==23.0.0
yookassa==3.3.0
